% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{aggregate_volsD}
\alias{aggregate_volsD}
\title{Recompute volumes for non-leaf nodes}
\usage{
aggregate_volsD(tree)
}
\arguments{
\item{tree}{the tree of interest}
}
\value{
the tree invisibly
}
\description{
Destructively aggregate volumes from the leaves to the
root. This sets each non-leaf node to be the sum of its
children
}
