% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{child_of}
\alias{child_of}
\title{Check parent}
\usage{
child_of(node, name)
}
\arguments{
\item{node}{The node of interest}

\item{name}{The name to look for in the nodes
ancestry}
}
\value{
boolean whether one of nodes ancestors
has the specified `name`
}
\description{
Check if node descends from node "name"
}
