% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compile_models}
\alias{compile_models}
\title{Compile a tree model}
\usage{
compile_models(model = c("effect_diffusion_tree", "simple_hierarchical"))
}
\arguments{
\item{model}{The model to compile, see the
default args for a list of available models.}
}
\value{
A \link{stan_model} object
}
\description{
Compile one of the models available in this
package.
}
