% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{compute_marg_lik}
\alias{compute_marg_lik}
\title{Compute marginal information criteria for some bayesian linear models}
\usage{
compute_marg_lik(model, data, marginalize, quadrature_nodes = 11, ...)
}
\arguments{
\item{model}{The model result (stanfit, stanreg)}

\item{data}{the input data used to fit the model (if not included in the model
object.}

\item{marginalize}{the random effect to marginalize over}

\item{quadrature_nodes}{the number of nodes to integrate over}

\item{...}{extra parameters for methods}
}
\description{
Compute marginal information criteria for some bayesian linear models
}
