% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_fit_and_pred.R
\name{compute_subject_error}
\alias{compute_subject_error}
\title{Compute subject specific sum of squares}
\usage{
compute_subject_error(pred, mod)
}
\arguments{
\item{The}{predicted value}

\item{The}{model of interest}
}
\description{
Compute subject specific sum of squares
}
