% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{create_tree}
\alias{create_tree}
\title{Generate a simulated tree}
\usage{
create_tree(noise_sd, leaf_effects, leaf_volumes, indiv_effects, metadata,
  base_tree, intercept_sd = 0, extra_effects = NA)
}
\arguments{
\item{noise_sd}{What should the sd of the gaussian noise
added to observations be}

\item{leaf_effects}{A named vector of effects to apply to
the leaves}

\item{indiv_effects}{A vector of individual level effects}

\item{metadata}{A vector containing a covariate
with elements matching the elements of the indiv effects / tree
volumes.}

\item{base_tree}{A tree indicating the ontology, should contain
volumes already to be overwritten.}

\item{extra_effects}{A named vector of extra effects to be applied
at nodes. Default \code{NA} for no extra effects}

\item{leaf_volume}{A named vector of scaling constants to
be applied to the leaf nodes, converting them to \code{volume} scale.}
}
\value{
A two element list containing
\enumerate{
\item The simulated tree
\item A vector with the composite \code{leaf_effects} generated by adding
\code{extra_effects} to \code{leaf_effects} where applicable.
}
}
\description{
Create a simulated tree from a noise sd,
named vectors of leaf effects, volumes, individual
effects, metadata, and an initial tree
}
