% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{edt}
\alias{edt}
\title{Fit the effect diffusion tree}
\usage{
edt(data, cores = 4)
}
\arguments{
\item{data}{A data object generated by \code{tree_to_edt_data}}

\item{cores}{number of cores to use, default 4}
}
\description{
Fits the effect diffusion tree model
}
