% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{effect_areas}
\alias{effect_areas}
\title{Effect areas}
\usage{
effect_areas(results, effects, sds, tree)
}
\arguments{
\item{results}{The result object of interest}

\item{effects}{The simulated effects}

\item{sds}{The node sds}

\item{tree}{The source tree}
}
\value{
a ggplot object
}
\description{
Create a posterior plot for the parameters with true effects
from results generated by \code{get_sglm_results} and \code{get_ept_results}
}
