% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_fit_and_pred.R
\name{extract_y}
\alias{extract_y}
\title{Get the y value for a model, useful in the case of no-pooling
where the variables get shuffled.}
\usage{
extract_y(mod)
}
\arguments{
\item{mod}{The model of interest}
}
\description{
Get the y value for a model, useful in the case of no-pooling
where the variables get shuffled.
}
