% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{fitted_ept}
\alias{fitted_ept}
\title{Get the fitted value from an ept model}
\usage{
fitted_ept(ept)
}
\arguments{
\item{ept}{The ept model of interest}
}
\value{
A n-vector of median fitted values
}
\description{
Get the fitted value from an ept model
}
