% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{fitted_np}
\alias{fitted_np}
\title{Get the fitted values from a list of unpooled models}
\usage{
fitted_np(np)
}
\arguments{
\item{np}{The data.frame of n no-pooling models}
}
\value{
A n-vector of median fitted_values
}
\description{
Get the fitted values from a list of unpooled models
}
