% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_names_and_plot}
\alias{fix_names_and_plot}
\title{Fix node names and plot a tree}
\usage{
fix_names_and_plot(tree)
}
\arguments{
\item{tree}{A tree to plot}
}
\value{
NULL invisibly
}
\description{
Fix node names then plot the tree
}
