% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_requirements}
\alias{get_data_requirements}
\title{Check the data requirements for a stan model}
\usage{
get_data_requirements(model)
}
\arguments{
\item{model}{The stan model to inspect}
}
\value{
a character vector of required data
}
\description{
Parse out the data block from a stan model and
report the required data
}
