% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{get_ept_results}
\alias{get_ept_results}
\title{Extract the salient results from an ept mod}
\usage{
get_ept_results(ept_mod, tree, sds, justLeaves = FALSE, addFix = TRUE)
}
\arguments{
\item{ept_mod}{The ept_mod}

\item{tree}{The tree of interest}

\item{sds}{SDs used in scaling leaf volumes}

\item{justLeaves}{Whether or not to just extract effects at leaves}

\item{addFix}{Whether to add the fixed effects before scaling}
}
\value{
A list containing
\enumerate{
\item \code{fix} The fixed effects of the model
\item \code{effects} The estimated median effects in the model
\item \code{ranints} The estimated random intercepts from the model
\item \code{b_post} The effect posterior samples
}
}
\description{
Extract the salient results from an ept mod
}
