% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_fit_and_pred.R
\name{get_fitted}
\alias{get_fitted}
\title{Get the fitted values from bayesian hierarchical linear models}
\usage{
get_fitted(mod)
}
\arguments{
\item{mod}{The model}
}
\description{
Much faster for gaussian linear models than the rstanarm equivalent
}
