% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{get_np_results}
\alias{get_np_results}
\title{Extract the salient results from an stan_glmer model}
\usage{
get_np_results(smod, tree, sds, justLeaves = FALSE, main = "SEXM")
}
\arguments{
\item{smod}{The \link{stan_glmer} model}

\item{tree}{The tree of interest}

\item{sds}{SDs used in scaling leaf volumes}

\item{justLeaves}{Whether or not to just extract effects at leaves}
}
\value{
A list containing
\enumerate{
\item \code{fix} 0 for this model
\item \code{effects} The estimated median effects in the model
\item \code{ranints} 0 for this model
\item \code{b_post} The posterior for the effects
}
}
\description{
Extract the salient results from an stan_glmer model
}
