% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{get_sglm_results}
\alias{get_sglm_results}
\title{Extract the salient results from an stan_glmer model}
\usage{
get_sglm_results(smod, tree, sds, justLeaves = FALSE)
}
\arguments{
\item{smod}{The \link{stan_glmer} model}

\item{tree}{The tree of interest}

\item{justLeaves}{Whether or not to just extract effects at leaves}

\item{leaf_sds}{SDs used in scaling leaf volumes}
}
\value{
A list containing
\enumerate{
\item \code{fix} The fixed effects of the model
\item \code{effects} The estimated median effects in the model
\item \code{ranints} The estimated random intercepts from the model
\item \code{b_post} The effect posterior samples
}
}
\description{
Extract the salient results from an stan_glmer model
}
