% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{h0}
\alias{h0}
\title{Fit a flat hierarchical model}
\usage{
h0(data, cores = 4)
}
\arguments{
\item{data}{A data object generated by \code{tree_to_edt_data}}

\item{cores}{number of cores to use, default 4}
}
\description{
Fits \code{scaled_vol ~ group + (group | p0) + (1 | ID)}
with \link{stan_lmer}
}
