% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{logLik_ept}
\alias{logLik_ept}
\title{Compute the likelihood for a tree model}
\usage{
logLik_ept(mod, y)
}
\arguments{
\item{mod}{A stan model with at least \code{y_pred}
and \code{sigma_model}}

\item{y}{The y data}
}
\value{
A matrix with the same dimensions as
\code{extract(mod, "y_pred")[[1]]} i.e. samples x obs
containing the gaussian likelihood of each
y_pred sample.
}
\description{
Compute the likelihood for a tree model
}
