% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{logLik_np}
\alias{logLik_np}
\title{Compute the likelihood for a tree model}
\usage{
logLik_np(mod)
}
\arguments{
\item{mod}{A data.frame with unpooled models}
}
\value{
A log-likelihood matrix
}
\description{
Compute the likelihood for a tree model
}
