% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{logLik_sglm_nocluster}
\alias{logLik_sglm_nocluster}
\title{Compute the marginal likelihood of a data point}
\usage{
logLik_sglm_nocluster(mod, y, indiv)
}
\arguments{
\item{y}{the observed data}

\item{indiv}{a character vector index to the individual}

\item{The}{model}
}
\value{
A matrix with the same dimensions as
\code{extract(mod, "y_pred")[[1]]} i.e. samples x obs
containing the gaussian likelihood of each
y_pred sample.

[[1]: R:[1
}
\description{
This assumes no knowledge of the individual random intercept from
and simulates unobserved subjects.
}
