% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{node_numbers}
\alias{node_numbers}
\title{Get node numbers}
\usage{
node_numbers(x, tree)
}
\arguments{
\item{x}{a vector of node names}

\item{tree}{a tree to get the index from}
}
\value{
a vector with length x containing the node indices
}
\description{
Take a vector of node names and return their index according
to a tree walked in pre-order.
}
