% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{np}
\alias{np}
\title{Fit the no-pooling model}
\usage{
np(data, cores = 1)
}
\arguments{
\item{data}{A data object generated by \code{tree_to_edt_data}}

\item{cores}{number of cores to use, default 4}
}
\description{
Fits a massively univariate model using \link{stan_glm}
with \code{scaled_vol ~ group}. This runs faster single core.
}
