% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{parent_index}
\alias{parent_index}
\title{Find the parent indices for each node}
\usage{
parent_index(tree)
}
\arguments{
\item{tree}{The tree of interest}
}
\value{
An integer vector with parent indices
}
\description{
Walks the tree in pre-order extracting the index
of the parent for each structure.
}
