% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_fit_and_pred.R
\name{post_pred}
\alias{post_pred}
\title{Get posterior predictions from a bayesian hierarchical linear model}
\usage{
post_pred(mod)
}
\arguments{
\item{mod}{The model}
}
\description{
Faster than the rstanarm version.
}
