% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{pw_effect_loglik}
\alias{pw_effect_loglik}
\title{Effect likelihoods}
\usage{
pw_effect_loglik(results, effects, sds, tree)
}
\arguments{
\item{results}{The result object of interest}

\item{effects}{The simulated effects}

\item{sds}{The node sds}

\item{tree}{The source tree}
}
\value{
A vector of pointwise likelihoods for each parameter
}
\description{
Compute the likelihood of the simulated parameters given the models
}
