% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{requirements_to_skeleton}
\alias{requirements_to_skeleton}
\title{Create a data skeleton for a model}
\usage{
requirements_to_skeleton(reqs)
}
\arguments{
\item{req}{The data requirements table}
}
\value{
A list with length equal to the number of rows
in \code{reqs} with names equal to the \code{name} column of
\code{reqs} and values initialized to either the contents
of the \code{value} column if it exists or NAs.
}
\description{
Take the results of \link{get_data_requirements}
and produce a skeleton list object to be passed to
stan. If the requirements table has a value column
it will be include as the contents of the list. Otherwise
list elements are initialized to NA.
}
