% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{set_leavesD}
\alias{set_leavesD}
\title{Set leaves}
\usage{
set_leavesD(tree, vals)
}
\arguments{
\item{tree}{the tree of interest}

\item{vals}{a scalar or vector of replacement values}
}
\value{
the tree invisibly
}
\description{
Set the volumes at the leaves of a tree to a specific value.
}
