% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{summarize_model}
\alias{summarize_model}
\title{Summarize a bayesian linear model}
\usage{
summarize_model(model, tree, leaf_scales = NULL, just_leaves = TRUE, ...)
}
\arguments{
\item{model}{the model to summarize}

\item{tree}{the tree used to fit the model}

\item{leaf_scales}{scaling factors for the leaves, useful
if the leaf response variables were scaled before modelling.}

\item{add_fixed_effect}{whether or not to incorporate the fixed
add the main effect to the leaf effects.}

\item{p0}{the leaf coefficient name}

\item{p1}{the first layer parent name}

\item{p2}{the second layer parent name}
}
\value{
A list containing
\enumerate{
\item \code{fix} The fixed effects of the model
\item \code{effects} The estimated median effects in the model
\item \code{ranints} The estimated random intercepts from the model
\item \code{b_post} The effect posterior samples
}
}
\description{
Extracts the salient results from a tree model. The new interface to
the old "get_*_results" functions
}
