% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{tag_volume_frame}
\alias{tag_volume_frame}
\title{Associate metadata with a volume tree frame}
\usage{
tag_volume_frame(vol_frame, metadata)
}
\arguments{
\item{vol_frame}{the volume frame}

\item{metadata}{the metadata frame}
}
\value{
the tagged data frame
}
\description{
Take a volume tree frame generated by [tree_to_volume_frame]
and associated metadata. The rows of metadata should correspond
to the order of the volumes in the original volume tree.
}
