% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{tree_to_edt_data}
\alias{tree_to_edt_data}
\alias{tree_to_ept_data_new}
\title{Convert a tree and metadata into useable data for hierarchical
modelling}
\usage{
tree_to_edt_data(tree, metadata, scale = FALSE, justLeaves = TRUE,
  model_formula = ~group)

tree_to_ept_data_new(tree, metadata, scale = FALSE, justLeaves = TRUE,
  model_formula = ~group)
}
\arguments{
\item{tree}{The tree containing volumes}

\item{metadata}{The metadata necessary, requires \code{SEX} as
the covariate.}

\item{scale}{Whether or not to centre/scale the volumes}

\item{justLeaves}{Whether to filter down the data to
just the leaf volume (default true)}
}
\value{
a list containing the requisite data for
fitting either flat models or effect diffusion models.
}
\description{
Convert a tree and metadata into useable data for hierarchical
modelling
}
\section{Functions}{
\itemize{
\item \code{tree_to_ept_data_new}: deprecated old variant
}}

