% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{tree_to_ept_data}
\alias{tree_to_ept_data}
\title{Convert a tree and metadata into useable data for ept
data}
\usage{
tree_to_ept_data(tree, metadata, justLeaves = FALSE)
}
\arguments{
\item{tree}{The tree containing volumes}

\item{metadata}{The metadata necessary, requires \code{SEX} as
the covariate.}
}
\value{
a list containing the requiste data for fitting an
ept model.
}
\description{
Convert a tree and metadata into useable data for ept
data
}
