% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{tree_to_volume_frame_new}
\alias{tree_to_volume_frame_new}
\title{Convert a volume tree to a data.frame}
\usage{
tree_to_volume_frame_new(vol_tree, filterFun = isLeaf)
}
\arguments{
\item{vol_tree}{The volume tree of interest}

\item{filterFun}{which nodes to include in the tree, defaults
to isLeaf, all can be included with \code{function(n) TRUE}}
}
\value{
A \code{data_frame} with the volume tree
}
\description{
Extract the name, label, and volumes for each node of the
tree. The volumes are then expanded out creating a
\code{data_frame} with n x v rows where n is the number of
rows and v is the number of volumes
}
