% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{add_effectD}
\alias{add_effectD}
\title{Increment volumes}
\usage{
add_effectD(tree, effect, name = tree$name)
}
\arguments{
\item{tree}{The tree of interest}

\item{effect}{a scalar or vector corresponding to the
amount to add to each leaf.}

\item{name}{The node name to increment at or below}
}
\value{
the tree invisibly
}
\description{
Destructively add effect to the volume
of all leaves at or below node "name"
}
