% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{scale_volumesD}
\alias{scale_volumesD}
\title{Scale node volumes up}
\usage{
scale_volumesD(tree, scale, name = tree$name)
}
\arguments{
\item{tree}{the tree of interest}

\item{scale}{a scalar or vector of how much to scale each
volume}

\item{name}{the node name to increment at or below}
}
\value{
the tree invisibly
}
\description{
Destructively apply a scale factor to the volumes of all leaf
nodes at or below node "name"
}
